Feature/sign v3

+++ contracts/OrionProxy.sol
@@ -1,27 +1,26 @@
 pragma solidity 0.5.10;
 
-import '@openzeppelin/contracts/ownership/Ownable.sol';
-import './Utils.sol';
-import './Validators/ValidatorV1.sol';
+import "@openzeppelin/contracts/ownership/Ownable.sol";
+import "./Utils.sol";
 
 /**
  * @title Proxy Contract
  * @dev Proxy Exchange contract for the Orion Protocol
  * @author @wafflemakr
  */
-contract OrionProxy is Ownable{
-
-    bytes32 public constant TYPE_HASH = 0x780982dd45b7930f3e71393eb3867ca735e735c553a8067145363bb3b7e2c47c;
+contract OrionProxy is Ownable {
+    bytes32
+        public constant TYPE_HASH = 0x780982dd45b7930f3e71393eb3867ca735e735c553a8067145363bb3b7e2c47c;
 
     uint256 private _guardCounter;
 
     enum Status {NEW, PARTIALLY_FILLED, FILLED, PARTIALLY_CANCELLED, CANCELLED}
 
-    struct Trade{
-        uint filledPrice;
-        uint filledAmount;
-        uint feePaid;
-        uint timestamp;
+    struct Trade {
+        uint256 filledPrice;
+        uint256 filledAmount;
+        uint256 feePaid;
+        uint256 timestamp;
     }
 
     // Get trades by orderHash
@@ -31,23 +30,22 @@ contract OrionProxy is Ownable{
     mapping(bytes32 => Status) public orderStatus;
 
     // Get user balance by address and asset address
-    mapping(address => mapping(address => uint)) private assetBalances;
+    mapping(address => mapping(address => uint256)) private assetBalances;
 
     // Pause or unpause exchangebuyOrderHash
     bool public isActive = true;
 
     address public implementation;
-    
 
-    constructor(address _implementation) public{
+    constructor(address _implementation) public {
         setImplementation(_implementation);
     }
 
-    function setImplementation(address _implementation) public onlyOwner{
+    function setImplementation(address _implementation) public onlyOwner {
         implementation = _implementation;
     }
 
-    function() external payable{
+    function() external payable {
         address _imp = implementation;
 
         assembly {
@@ -65,10 +63,12 @@ contract OrionProxy is Ownable{
 
             // (4) forward return data back to caller
             switch result
-            case 0 { revert(ptr, size) }
-            default { return(ptr, size) }
+                case 0 {
+                    revert(ptr, size)
+                }
+                default {
+                    return(ptr, size)
+                }
         }
-
     }
-
 }
